transcriber3000/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==1.10.3
beautifulsoup4==4.8.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
compressed_rtf==1.0.6
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.2.2
extract-msg==0.28.7
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.4
jiter==0.8.0
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
olefile==0.47
openai==1.55.3
openai-whisper==20240930
pdfminer.six==20191110
pillow==11.0.0
pycryptodome==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
python-pptx==0.6.23
regex==2024.11.6
requests==2.32.3
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
xlrd==1.2.0
XlsxWriter==3.2.0
